Split large 'next-routes' into a number of smaller files
