+++ jogging-app/local_modules/next-routes/utils/perform-redirect.js
@@ -1,0 +1,8 @@
+export const performRedirect = ({ req, res, redirectTo }) => {
+    const { baseUrl } = req.appMeta;
+
+    res.writeHead(301, {
+        Location: baseUrl + redirectTo
+    });
+    res.end();
+};
