+++ jogging-app/local_modules/next-routes/utils/to-query-string.js
@@ -1,0 +1,15 @@
+export const toQueryString = obj =>
+    Object.keys(obj)
+        .filter(key => obj[key] !== null && obj[key] !== undefined)
+        .map(key => {
+            let value = obj[key];
+
+            if (Array.isArray(value)) {
+                value = value.join('/');
+            }
+
+            return [encodeURIComponent(key), encodeURIComponent(value)].join(
+                '='
+            );
+        })
+        .join('&');
